\name{SA23204177-package}
\alias{SA23204177-package}
\alias{SA23204177}
\docType{package}
\title{
  Test the mean of highdimensional samples
}
\description{
  Testing whether the sample mean vector is a zero vector using R, Rcpp and RcppEigen. 
}
\author{
Mingming Xiang

Maintainer: Mingming Xiang <xmm483754@gmail.com>
}
\references{
  Runmin Wang, Xiaofeng Shao. (2020) "Hypothesis testing for high-dimensional time series via self-normalization." Ann. Statist. 48 (5) 2728 - 2758, October 2020. https://doi.org/10.1214/19-AOS1904
}
\keyword{ package }
