% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Wang2020C}
\alias{Wang2020C}
\title{Testing of mean vector using Rcpp and RcppEigen}
\usage{
Wang2020C(sam, m, n)
}
\arguments{
\item{sam}{the matrix of samples}

\item{m}{trimming parameter}

\item{n}{sample size}
}
\value{
\code{TRUE} (reject null hypothesis) or \code{FALSE}
}
\description{
Testing of mean vector using Rcpp and RcppEigen
}
\examples{
\dontrun{
library(MASS)
set.seed(1234)
n1 <- n2 <- 50
p <- 200
mu1 <- rep(0, p)
mu2 <- mu1
mu2[1:10] <- 1
true.cov <- 0.4^(abs(outer(1:p, 1:p, "-"))) # AR1 covariance
sam1 <- mvrnorm(n = n1, mu = mu1, Sigma = true.cov)
sam2 <- mvrnorm(n = n2, mu = mu2, Sigma = true.cov)
Wang2020C(sam1, m=5, n=n1)
Wang2020C(sam2, m=5, n=n2)
}
}
